Always set the cursor position
Fix InsertHeader leaves you at the bottom of the file when inserting the
first include.

Switch from getpos to getcurpos since that should more accurately get
our position.

When we don't find an include, we'll be stuck in the start/end of the
file (where we went with normal). Instead, position the cursor to where
the include was inserted.